+++ com.google.android.appbundle/Editor/Scripts/AssetPack.cs
@@ -29,6 +29,8 @@ namespace Google.Android.AppBundle.Editor
         private Dictionary<TextureCompressionFormat, string> _compressionFormatToAssetPackDirectoryPath;
         private Dictionary<DeviceTier, string> _deviceTierToAssetBundleFilePath;
         private Dictionary<DeviceTier, string> _deviceTierToAssetPackDirectoryPath;
+        private Dictionary<string, string> _deviceGroupToAssetBundleFilePath;
+        private Dictionary<string, string> _deviceGroupToAssetPackDirectoryPath;
 
 
         /// <summary>
@@ -79,6 +81,15 @@ namespace Google.Android.AppBundle.Editor
                     throw new ArgumentException("DeviceTierToAssetPackDirectoryPath is already set.");
                 }
 
+                if (_deviceGroupToAssetBundleFilePath != null)
+                {
+                    throw new ArgumentException("DeviceGroupToAssetBundleFilePath is already set.");
+                }
+
+                if (_deviceGroupToAssetPackDirectoryPath != null)
+                {
+                    throw new ArgumentException("DeviceGroupToAssetPackDirectoryPath is already set.");
+                }
 
                 _assetBundleFilePath = value;
             }
@@ -126,6 +137,15 @@ namespace Google.Android.AppBundle.Editor
                     throw new ArgumentException("DeviceTierToAssetPackDirectoryPath is already set.");
                 }
 
+                if (_deviceGroupToAssetBundleFilePath != null)
+                {
+                    throw new ArgumentException("DeviceGroupToAssetBundleFilePath is already set.");
+                }
+
+                if (_deviceGroupToAssetPackDirectoryPath != null)
+                {
+                    throw new ArgumentException("DeviceGroupToAssetPackDirectoryPath is already set.");
+                }
 
                 _assetPackDirectoryPath = value;
             }
@@ -176,6 +196,15 @@ namespace Google.Android.AppBundle.Editor
                     throw new ArgumentException("DeviceTierToAssetPackDirectoryPath is already set.");
                 }
 
+                if (_deviceGroupToAssetBundleFilePath != null)
+                {
+                    throw new ArgumentException("DeviceGroupToAssetBundleFilePath is already set.");
+                }
+
+                if (_deviceGroupToAssetPackDirectoryPath != null)
+                {
+                    throw new ArgumentException("DeviceGroupToAssetPackDirectoryPath is already set.");
+                }
 
                 _compressionFormatToAssetBundleFilePath = value;
             }
@@ -226,6 +255,15 @@ namespace Google.Android.AppBundle.Editor
                     throw new ArgumentException("DeviceTierToAssetPackDirectoryPath is already set.");
                 }
 
+                if (_deviceGroupToAssetBundleFilePath != null)
+                {
+                    throw new ArgumentException("DeviceGroupToAssetBundleFilePath is already set.");
+                }
+
+                if (_deviceGroupToAssetPackDirectoryPath != null)
+                {
+                    throw new ArgumentException("DeviceGroupToAssetPackDirectoryPath is already set.");
+                }
 
                 _compressionFormatToAssetPackDirectoryPath = value;
             }
@@ -273,6 +311,16 @@ namespace Google.Android.AppBundle.Editor
                 if (_deviceTierToAssetPackDirectoryPath != null)
                 {
                     throw new ArgumentException("DeviceTierToAssetPackDirectoryPath is already set.");
+                }
+
+                if (_deviceGroupToAssetBundleFilePath != null)
+                {
+                    throw new ArgumentException("DeviceGroupToAssetBundleFilePath is already set.");
+                }
+
+                if (_deviceGroupToAssetPackDirectoryPath != null)
+                {
+                    throw new ArgumentException("DeviceGroupToAssetPackDirectoryPath is already set.");
                 }
 
                 _deviceTierToAssetBundleFilePath = value;
@@ -324,7 +372,133 @@ namespace Google.Android.AppBundle.Editor
                     throw new ArgumentException("DeviceTierToAssetBundleFilePath is already set.");
                 }
 
+                if (_deviceGroupToAssetBundleFilePath != null)
+                {
+                    throw new ArgumentException("DeviceGroupToAssetBundleFilePath is already set.");
+                }
+
+                if (_deviceGroupToAssetPackDirectoryPath != null)
+                {
+                    throw new ArgumentException("DeviceGroupToAssetPackDirectoryPath is already set.");
+                }
+
                 _deviceTierToAssetPackDirectoryPath = value;
+            }
+        }
+
+        /// <summary>
+        /// Dictionary from device group to the location on disk of a single AssetBundle file.
+        /// When using Play Asset Delivery APIs, only the AssetBundle for the device's group will be delivered.
+        /// </summary>
+        /// <exception cref="ArgumentException">
+        /// Thrown on set if any other path property is already set (non-null).
+        /// </exception>
+        public Dictionary<string, string> DeviceGroupToAssetBundleFilePath
+        {
+            get { return _deviceGroupToAssetBundleFilePath; }
+            set
+            {
+                if (value == null)
+                {
+                    _deviceGroupToAssetBundleFilePath = null;
+                    return;
+                }
+
+                if (_assetBundleFilePath != null)
+                {
+                    throw new ArgumentException("AssetBundleFilePath is already set.");
+                }
+
+                if (_assetPackDirectoryPath != null)
+                {
+                    throw new ArgumentException("AssetPackDirectoryPath is already set.");
+                }
+
+                if (_compressionFormatToAssetBundleFilePath != null)
+                {
+                    throw new ArgumentException("CompressionFormatToAssetBundleFilePath is already set.");
+                }
+
+                if (_compressionFormatToAssetPackDirectoryPath != null)
+                {
+                    throw new ArgumentException("CompressionFormatToAssetPackDirectoryPath is already set.");
+                }
+
+
+                if (_deviceTierToAssetBundleFilePath != null)
+                {
+                    throw new ArgumentException("DeviceTierToAssetBundleFilePath is already set.");
+                }
+
+                if (_deviceTierToAssetPackDirectoryPath != null)
+                {
+                    throw new ArgumentException("DeviceTierToAssetPackDirectoryPath is already set.");
+                }
+
+                if (_deviceGroupToAssetPackDirectoryPath != null)
+                {
+                    throw new ArgumentException("DeviceGroupToAssetPackDirectoryPath is already set.");
+                }
+
+                _deviceGroupToAssetBundleFilePath = value;
+            }
+        }
+
+        /// <summary>
+        /// Dictionary from device group to the location on disk of a folder containing
+        /// raw asset files.
+        /// When using Play Asset Delivery APIs, only the folder for the device's group will be delivered.
+        /// </summary>
+        /// <exception cref="ArgumentException">
+        /// Thrown on set if any other path property is already set (non-null).
+        /// </exception>
+        public Dictionary<string, string> DeviceGroupToAssetPackDirectoryPath
+        {
+            get { return _deviceGroupToAssetPackDirectoryPath; }
+            set
+            {
+                if (value == null)
+                {
+                    _deviceGroupToAssetPackDirectoryPath = null;
+                    return;
+                }
+
+                if (_assetBundleFilePath != null)
+                {
+                    throw new ArgumentException("AssetBundleFilePath is already set.");
+                }
+
+                if (_assetPackDirectoryPath != null)
+                {
+                    throw new ArgumentException("AssetPackDirectoryPath is already set.");
+                }
+
+                if (_compressionFormatToAssetBundleFilePath != null)
+                {
+                    throw new ArgumentException("CompressionFormatToAssetBundleFilePath is already set.");
+                }
+
+                if (_compressionFormatToAssetPackDirectoryPath != null)
+                {
+                    throw new ArgumentException("CompressionFormatToAssetPackDirectoryPath is already set.");
+                }
+
+                if (_deviceTierToAssetBundleFilePath != null)
+                {
+                    throw new ArgumentException("DeviceTierToAssetBundleFilePath is already set.");
+                }
+
+                if (_deviceTierToAssetPackDirectoryPath != null)
+                {
+                    throw new ArgumentException("DeviceTierToAssetPackDirectoryPath is already set.");
+                }
+
+                if (_deviceGroupToAssetBundleFilePath != null)
+                {
+                    throw new ArgumentException("DeviceGroupToAssetBundleFilePath is already set.");
+                }
+
+                _deviceGroupToAssetPackDirectoryPath = value;
             }
         }