+++ com.google.android.appbundle/Editor/Scripts/Internal/BuildTools/BundletoolConfig.cs
@@ -34,6 +34,7 @@ namespace Google.Android.AppBundle.Editor.Internal.BuildTools
         public const string ScreenDensity = "SCREEN_DENSITY";
         public const string TextureCompressionFormat = "TEXTURE_COMPRESSION_FORMAT";
         public const string DeviceTier = "DEVICE_TIER";
+        public const string DeviceGroup = "DEVICE_GROUP";
 
 
         // Options for BundleType